Update Book & Cart

+++ server/model/CartModel.php
@@ -1,5 +1,5 @@
 <?php 
-    require_once __DIR__ . '/../connect.php/';
+    require_once __DIR__ . '/../connect.php';
 
     class CartModel {
         private $con;
@@ -25,6 +25,7 @@
             if ($existed) {
                 $newQuantity = $existed['quantity'] + $quantity;
                 $tmp = array(
+                    'user_id' => $user_id,
                     'book_id' => $book_id,
                     'quantity' => $newQuantity
                 );
@@ -42,11 +43,12 @@
         }
 
         // input: user_id, book_id
-        // output: list of item in matched user's cart
+        // output: a matched item
         public function read($params) {
-            $user_id = intval($params['book_id']);
+            $user_id = intval($params['user_id']);
+            $book_id = intval($params['book_id']);
 
-            $query = "SELECT * FROM CART WHERE user_id = $user_id";
+            $query = "SELECT * FROM CART WHERE user_id = $user_id AND book_id = $book_id";
             $result = mysqli_query($this->con, $query);
 
             $items = null;
@@ -59,22 +61,26 @@
             return $items;
         }
 
-        // input: book_id, quantity
+        // input: user_id, book_id, quantity
         // output: bool
         public function update($params) {
+            $user_id = intval($params['user_id']);
             $book_id = intval($params['book_id']);
             $quantity = intval($params['quantity']);
 
-            $query = "UPDATE CART SET quantity = $quantity WHERE book_id = $book_id";
+            $query = "UPDATE CART SET quantity = $quantity WHERE user_id = $user_id AND book_id = $book_id";
             $result = mysqli_query($this->con, $query);
 
             return $result ? true : false;
         }
 
-        public function delete($book_id) {
-            $book_id = intval($book_id);
+        // intput: user_id, book_id
+        // output: bool
+        public function delete($params) {
+            $user_id = intval($params['user_id']);
+            $book_id = intval($params['book_id']);
 
-            $query = "DELETE FROM CART WHERE book_id = $book_id";
+            $query = "DELETE FROM CART WHERE user_id = $user_id AND book_id = $book_id";
             $result = mysqli_query($this->con, $query);
 
             return $result ? true : false;
